Rename forms to PascalCase (#9)
